db, state: migrate db types to state
