+++ consensus/babe/babe_test.go
@@ -284,7 +284,7 @@ func createFlatBlockTree(t *testing.T, depth int) *blocktree.BlockTree {
 	}
 	genesisBlock.SetBlockArrivalTime(uint64(1000))
 
-	d := &db.BlockDB{
+	d := &state.BlockDB{
 		Db: db.NewMemDatabase(),
 	}
